Types fix: librandom, libstdcrypto

+++ src/include/skalibs/random.h
@@ -3,7 +3,8 @@
 #ifndef RANDOM_H
 #define RANDOM_H
 
-#include <skalibs/uint32.h>
+#include <sys/types.h>
+#include <stdint.h>
 #include <skalibs/stralloc.h>
 
 extern void random_makeseed (char *) ; /* fills 160 bytes */
@@ -12,10 +13,10 @@ extern int random_init (void) ;
 extern void random_finish (void) ;
 
 extern unsigned char random_char (void) ;
-extern void random_string (char *, unsigned int) ;
-extern uint32 random_uint32 (uint32) ;
-extern void random_name (char *, unsigned int) ;
-extern void random_unsort (char *, unsigned int, unsigned int) ;
-extern int random_sauniquename (stralloc *, unsigned int) ;
+extern void random_string (char *, size_t) ;
+extern uint32_t random_uint32 (uint32_t) ;
+extern void random_name (char *, size_t) ;
+extern void random_unsort (char *, size_t, size_t) ;
+extern int random_sauniquename (stralloc *, size_t) ;
 
 #endif
