+++ src/include/skalibs/surf.h
@@ -3,13 +3,14 @@
 #ifndef SKALIBS_SURF_H
 #define SKALIBS_SURF_H
 
-#include <skalibs/uint32.h>
+#include <sys/types.h>
+#include <stdint.h>
 
 typedef struct SURFSchedule SURFSchedule, *SURFSchedule_ref, **SURFSchedule_ref_ref ;
 struct SURFSchedule
 {
-  uint32 seed[32] ;
-  uint32 in[12] ;
+  uint32_t seed[32] ;
+  uint32_t in[12] ;
   char out[32] ;
   unsigned int pos ;
 } ;
@@ -17,6 +18,6 @@ struct SURFSchedule
 #define SURFSCHEDULE_ZERO { .seed = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 }, .in = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 }, .out = "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0", .pos = 32 }
 
 extern void surf_init (SURFSchedule *, char const *) ; /* 160 chars */
-extern void surf (SURFSchedule *, char *, unsigned int) ;
+extern void surf (SURFSchedule *, char *, size_t) ;
 
 #endif
