+++ src/librandom/random_name.c
@@ -4,7 +4,7 @@
 
 #include <skalibs/random.h>
 
-void random_name (char *s, unsigned int n)
+void random_name (char *s, size_t n)
 {
   static char const oklist[64] = "ABCDEFGHIJKLMNOPQRSTUVWXYZghijklmnopqrstuvwxyz-_0123456789abcdef" ;
   random_string(s, n) ;
