+++ src/libstdcrypto/md5_update.c
@@ -1,11 +1,13 @@
 /* ISC license. */
 
+#include <sys/types.h>
+#include <stdint.h>
 #include <skalibs/uint32.h>
 #include <skalibs/bytestr.h>
 #include <skalibs/md5.h>
 #include "md5-internal.h"
 
-void md5_update (MD5Schedule *ctx, char const *s, unsigned int len)
+void md5_update (MD5Schedule *ctx, char const *s, size_t len)
 {
   register uint32 t = ctx->bits[0] ;
   if ((ctx->bits[0] = t + (len << 3)) < t)
@@ -23,14 +25,14 @@ void md5_update (MD5Schedule *ctx, char const *s, unsigned int len)
     }
     byte_copy((char *)p, t, s) ;
     uint32_little_endian((char *)ctx->in, 16) ;
-    md5_transform(ctx->buf, (uint32 *)ctx->in) ;
+    md5_transform(ctx->buf, (uint32_t *)ctx->in) ;
     s += t ; len -= t ;
   }
   while (len >= 64)
   {
     byte_copy((char *)ctx->in, 64, s) ;
     uint32_little_endian((char *)ctx->in, 16) ;
-    md5_transform(ctx->buf, (uint32 *)ctx->in) ;
+    md5_transform(ctx->buf, (uint32_t *)ctx->in) ;
     s += 64 ; len -= 64 ;
   }
   byte_copy((char *)ctx->in, len, s) ;
